requirements.txt
scipy==1.2.1
tensorflow_gpu==1.8.0
numpy==1.19.2

